ci(agent): add workflow for building agent outside of a release (#4512)
* ci(agent): add workflow for building agent outside of a release

* ci(agent): skip signing for test workflow

* ci(agent): remove installation of java since it's not needed

+++ scripts/build-agent-installer-win64.sh
@@ -2,6 +2,9 @@
 
 # Agent Installer build script
 
+# Fail on error
+set -e
+
 # References:
 # https://docs.digicert.com/en/software-trust-manager/sign-with-digicert-signing-tools/sign-with-smctl.html
 # https://docs.digicert.com/en/software-trust-manager/ci-cd-integrations/script-integrations/github-actions-integration-with-pkcs11.html
@@ -10,11 +13,6 @@
 # https://github.com/ebourg/jsign
 
 # Pre-requisites
-
-if ! command -v java >/dev/null 2>&1; then
-    echo "Java required"
-    exit 1
-fi
 
 if ! command -v makensis >/dev/null 2>&1; then
     echo "makensis required"
@@ -26,36 +24,41 @@ if ! command -v wget >/dev/null 2>&1; then
     exit 1
 fi
 
-if [[ -z "${SM_HOST}" ]]; then
-  echo "DigiCert SM_HOST is missing"
-  exit 1
+# If we are skipping signing, none of the pre-reqs for signing are required
+if [ ! SKIP_SIGNING ]; then
+  if ! command -v java >/dev/null 2>&1; then
+      echo "Java required"
+      exit 1
+  fi
+
+  if [[ -z "${SM_HOST}" ]]; then
+    echo "DigiCert SM_HOST is missing"
+    exit 1
+  fi
+
+  if [[ -z "${SM_API_KEY}" ]]; then
+    echo "DigiCert SM_API_KEY is missing"
+    exit 1
+  fi
+
+  if [[ -z "${SM_CLIENT_CERT_FILE_BASE64}" ]]; then
+    echo "DigiCert SM_CLIENT_CERT_FILE_BASE64 is missing"
+    exit 1
+  fi
+
+  if [[ -z "${SM_CLIENT_CERT_PASSWORD}" ]]; then
+    echo "DigiCert SM_CLIENT_CERT_PASSWORD is missing"
+    exit 1
+  fi
+
+  if [[ -z "${SM_CERT_ALIAS}" ]]; then
+    echo "DigiCert SM_CERT_ALIAS is missing"
+    exit 1
+  fi
+
+  # Diagnostics
+  java --version
 fi
-
-if [[ -z "${SM_API_KEY}" ]]; then
-  echo "DigiCert SM_API_KEY is missing"
-  exit 1
-fi
-
-if [[ -z "${SM_CLIENT_CERT_FILE_BASE64}" ]]; then
-  echo "DigiCert SM_CLIENT_CERT_FILE_BASE64 is missing"
-  exit 1
-fi
-
-if [[ -z "${SM_CLIENT_CERT_PASSWORD}" ]]; then
-  echo "DigiCert SM_CLIENT_CERT_PASSWORD is missing"
-  exit 1
-fi
-
-if [[ -z "${SM_CERT_ALIAS}" ]]; then
-  echo "DigiCert SM_CERT_ALIAS is missing"
-  exit 1
-fi
-
-# Fail on error
-set -e
-
-# Diagnostics
-java --version
 
 # Get the current version number
 export MEDPLUM_VERSION=$(node -p "require('./package.json').version")
@@ -69,40 +72,42 @@ npm run build
 # Build the executable
 npx pkg ./dist/cjs/index.cjs --targets node18-win-x64 --output "dist/medplum-agent-$MEDPLUM_VERSION-win64.exe" --options no-warnings
 
-# Download JSign
-if [ ! -f "jsign-5.0.jar" ]; then
-    wget https://github.com/ebourg/jsign/releases/download/5.0/jsign-5.0.jar
+if [ ! SKIP_SIGNING ]; then
+  # Download JSign
+  if [ ! -f "jsign-5.0.jar" ]; then
+      wget https://github.com/ebourg/jsign/releases/download/5.0/jsign-5.0.jar
+  fi
+
+  # Unpack the client certificate
+  export SM_CLIENT_CERT_FILE="Certificate_pkcs12.p12"
+  echo "$SM_CLIENT_CERT_FILE_BASE64" | base64 --decode > "$SM_CLIENT_CERT_FILE"
+
+  # Sign the executable
+  java -jar jsign-5.0.jar \
+    --storetype DIGICERTONE \
+    --storepass "$SM_API_KEY|$SM_CLIENT_CERT_FILE|$SM_CLIENT_CERT_PASSWORD" \
+    --alias "$SM_CERT_ALIAS" \
+    "dist/medplum-agent-$MEDPLUM_VERSION-win64.exe"
+
+  # Download Shawl exe
+  rm -f shawl-v1.4.0-win64.zip
+  wget https://github.com/mtkennerly/shawl/releases/download/v1.4.0/shawl-v1.4.0-win64.zip
+  unzip shawl-v1.4.0-win64.zip
+  mv shawl.exe dist/shawl-v1.4.0-win64.exe
+
+  # Download Shawl legal
+  rm -f shawl-v1.4.0-legal.zip
+  wget https://github.com/mtkennerly/shawl/releases/download/v1.4.0/shawl-v1.4.0-legal.zip
+  unzip shawl-v1.4.0-legal.zip
+  mv shawl-v1.4.0-legal.txt dist
+
+  # Sign the Shawl executable
+  java -jar jsign-5.0.jar \
+    --storetype DIGICERTONE \
+    --storepass "$SM_API_KEY|$SM_CLIENT_CERT_FILE|$SM_CLIENT_CERT_PASSWORD" \
+    --alias "$SM_CERT_ALIAS" \
+    dist/shawl-v1.4.0-win64.exe
 fi
-
-# Unpack the client certificate
-export SM_CLIENT_CERT_FILE="Certificate_pkcs12.p12"
-echo "$SM_CLIENT_CERT_FILE_BASE64" | base64 --decode > "$SM_CLIENT_CERT_FILE"
-
-# Sign the executable
-java -jar jsign-5.0.jar \
-  --storetype DIGICERTONE \
-  --storepass "$SM_API_KEY|$SM_CLIENT_CERT_FILE|$SM_CLIENT_CERT_PASSWORD" \
-  --alias "$SM_CERT_ALIAS" \
-  "dist/medplum-agent-$MEDPLUM_VERSION-win64.exe"
-
-# Download Shawl exe
-rm -f shawl-v1.4.0-win64.zip
-wget https://github.com/mtkennerly/shawl/releases/download/v1.4.0/shawl-v1.4.0-win64.zip
-unzip shawl-v1.4.0-win64.zip
-mv shawl.exe dist/shawl-v1.4.0-win64.exe
-
-# Download Shawl legal
-rm -f shawl-v1.4.0-legal.zip
-wget https://github.com/mtkennerly/shawl/releases/download/v1.4.0/shawl-v1.4.0-legal.zip
-unzip shawl-v1.4.0-legal.zip
-mv shawl-v1.4.0-legal.txt dist
-
-# Sign the Shawl executable
-java -jar jsign-5.0.jar \
-  --storetype DIGICERTONE \
-  --storepass "$SM_API_KEY|$SM_CLIENT_CERT_FILE|$SM_CLIENT_CERT_PASSWORD" \
-  --alias "$SM_CERT_ALIAS" \
-  dist/shawl-v1.4.0-win64.exe
 
 # Build the installer
 makensis installer.nsi
